RS filtering tidying

+++ routes/web-auth.php
@@ -212,23 +212,25 @@ if( !config( 'ixp_fe.frontend.disabled.docstore_customer' ) ) {
 ///
 /// Rs Filtering
 ///
-Route::get( 'rs-filtering/{cust}', 'RsFilterController@list' )->name( 'rs-filter@list' );
+///
+if( !config( 'ixp_fe.frontend.disabled.rs-filters') ) {
+    Route::get( 'rs-filtering/{cust}', 'RsFilterController@list' )->name( 'rs-filter@list' );
 
-Route::group( [ 'prefix' => 'rs-filter' ], function() {
-    Route::get('create/{cust}',                  'RsFilterController@create'            )->name("rs-filter@create"             );
-    Route::get('edit/{rsf}',                     'RsFilterController@edit'           )->name("rs-filter@edit"            );
-    Route::get('view/{rsf}',                     'RsFilterController@view'           )->name("rs-filter@view"            );
-    Route::get('toogle-enable/{rsf}/{enable}',   'RsFilterController@toggleEnable'   )->name("rs-filter@toggle-enable"   );
-    Route::get('change-order/{rsf}/{up}',        'RsFilterController@changeOrderBy'  )->name("rs-filter@change-order"    );
+    Route::group( [ 'prefix' => 'rs-filter' ], function() {
+        Route::get('create/{cust}',                  'RsFilterController@create'            )->name("rs-filter@create"             );
+        Route::get('edit/{rsf}',                     'RsFilterController@edit'           )->name("rs-filter@edit"            );
+        Route::get('view/{rsf}',                     'RsFilterController@view'           )->name("rs-filter@view"            );
+        Route::get('toogle-enable/{rsf}/{enable}',   'RsFilterController@toggleEnable'   )->name("rs-filter@toggle-enable"   );
+        Route::get('change-order/{rsf}/{up}',        'RsFilterController@changeOrderBy'  )->name("rs-filter@change-order"    );
 
-    Route::post('revert/{cust}',                 'RsFilterController@revert'         )->name("rs-filter@revert"          );
-    Route::post('commit/{cust}',                 'RsFilterController@commit'         )->name("rs-filter@commit"          );
+        Route::post('revert/{cust}',                 'RsFilterController@revert'         )->name("rs-filter@revert"          );
+        Route::post('commit/{cust}',                 'RsFilterController@commit'         )->name("rs-filter@commit"          );
 
 
-    Route::post('store',                        'RsFilterController@store'          )->name("rs-filter@store"           );
-    Route::put('update/{rsf}',                  'RsFilterController@update'         )->name("rs-filter@update"          );
-    Route::delete('delete/{rsf}',               'RsFilterController@delete'       )->name("rs-filter@delete"          );
+        Route::post('store',                        'RsFilterController@store'          )->name("rs-filter@store"           );
+        Route::put('update/{rsf}',                  'RsFilterController@update'         )->name("rs-filter@update"          );
+        Route::delete('delete/{rsf}',               'RsFilterController@delete'       )->name("rs-filter@delete"          );
 
-    Route::view( 'grant-cust-user',              'rs-filter/grant-cust-user'        )->name( 'rs-filter@grant-cust-user' );
-});
-
+        Route::view( 'grant-cust-user',              'rs-filter/grant-cust-user'        )->name( 'rs-filter@grant-cust-user' );
+    });
+}
